Fixed errors with restructuring
