+++ src/core/navbar/index.js
@@ -1,8 +1,7 @@
-import React from 'react';
-import { connect } from 'react-redux';
 import Navbar from './navbar';
+import { connect } from 'react-redux';
 import { receiveQuery } from 'common/query/actions';
-import { loginUser, receiveUser } from 'common/modules/user/actions';
+import { loginUser, receiveUser } from 'modules/user/actions';
 
 const mapStateToProps = ({ user }) => ({
   user,
